Change a declaration so that it doesn't confuse C++ any more.
[Reviewed by Sung.]

This changes the chpl_qthread_comm_task_tls declaration I added in
r23192.  With the original declaration, when g++ compiled RE2 it made a
.bss symbol for that variable in re2-interface.o.  Unfortunately the C
compiler put the same variable in .data, and the mismatch resulted in
linker errors.  In this commit I declare chpl_qthread_comm_task_tls as
extern, which makes g++ produce a linkable symbol for it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23200 3a8e244f-b0f2-452b-bcba-4c88e055c3ca