Adjusted .bad for a future.
[not reviewed]

The failure mode for users/vass/crash1callDestructorsMain.chpl
changed upon r23156.

Before there was memory corruption - detectable with valgrind,
luckily for us we had a .bad and the failure was consistent.
It seems it was due to reusing a freed AST node.

Now we get this:
  ./crash1callDestructorsAddon.chpl:79: internal error: function is not normal [symbol.cpp:1911]

I am adjusting .bad and .future to reflect this change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23204 3a8e244f-b0f2-452b-bcba-4c88e055c3ca