Removing this .perfnumtrials file
Because stream-ep first takes the min across locales and then takes
the min/max/avg across those mins, when we're doing 1-locale
performance trials, we don't actually get stats for multiple
trials since we're talking the min/max/avg over a single value.
Plus, this test runs plenty fast.  For that reason, removing its
limit of 1 performance trial so that we get some spread in the
nightlies for a change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23206 3a8e244f-b0f2-452b-bcba-4c88e055c3ca