Add -Wno-inline more selectively than we did previously
[written by Elliot, reviewed by me]

In r21985, I added -Wno-inline to our list of compiler options used
when compiling in optimized (-O) mode because we use GASNet's flags,
it threw -Winline, and this generated warnings for some of our longer
routines at the time, which wasn't what we wanted.  Unfortunately, in
doing so, I wasn't careful to make sure that we only added it for
compilers that supported it, which broke gasnet compiles with -O for
compilers like PGI that don't know what to do with -Wno-inline.

This fix, from Elliot, only adds the flag for compilers that we know
support it.  An alternate way to approach this would be to add a
variable in the make/compiler/Makefile.* files on a
compiler-by-compiler basis that undoes this flag when necessary. We
already do this for the runtime for certain flags via the variable
CHPL_GASNET_MORE_CFLAGS.  However, we don't currently apply this
variable to generated code (perhaps we should?) so adding -Wno-inline
to it isn't sufficient as things stand today.  Or, perhaps we should
generate a second such variable just for generated code?  Or perhaps
the original variable should be moved into Makefile.comm-gasnet like
this since it is GASNet-specific to begin with.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23208 3a8e244f-b0f2-452b-bcba-4c88e055c3ca