Updated a spec example.
[not reviewed]

In the following example that we used to have:

-----------
var x: int;
cobegin {
  while x != 1 do ;  // spin wait
  x = 1;
}
-----------

the 'while' look is guaranteed not to terminate, and x=1 is illegal.

Added 'ref(x)' to the 'cobegin' to restore the spirit of the example
given our task intents.

The reason we were not getting an error for x=1 is that the entire
snippet was invisibly wrapped into an 'if false'.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23215 3a8e244f-b0f2-452b-bcba-4c88e055c3ca