Generate the "This compiler was built without LLVM support" error early.
[reviewed by Tom]

Prior to this change, the compiler built without LLVM support and
invoked with --llvm would go all the way into codegen then crash with an
internal error.

Instead, in this scenario we now will generate a user error right
after parsing the options.

There was no obvious place where to put this check, so I put it in
setupDependentVars(). Maybe the way to do it is to rename it to
postprocessArgs().


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23222 3a8e244f-b0f2-452b-bcba-4c88e055c3ca