Embed checkboxes in URL's querystring
[reviewed by tom]

This is so that what graphs you want to show can be embedded in the url to make
it easy to select the same graphs across different machines and to make it easy
to give somebody a pointer to particular graphs you want them to look at.  

This reads the query string on initialization and creates the graphs. The query
string is only updated after new graphs are displayed (as in if you just select
a check box it doesn't change the querystring. It only does that after you
click the display button or select a suite.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23226 3a8e244f-b0f2-452b-bcba-4c88e055c3ca