Don't sync the perf graphs to sourceforge if there were errors creating them
[reviewed by lydia] 

Since we're moving towards making the graphs public (and probably getting rid
of the internal copies) we want to make sure that the graphs are still
available even if the most recent graph generation failed. Currently we sync
the "graphs" no matter what but if there was a failure creating them we are
really just syncing over an empty folder. It seems better to just not sync on
failure and have day old data. 

To do this, genGraphs will create a SUCCESS file in the html directory if we
didn't detect any errors and all the supporting files were created. We use a
file and not an env var or something because genGraphs gets called in
start_test but the syncing happens in nightly so genGraphs couldn't set its
parent processes env var. 

The script that syncs to sourceforge now checks to see if the success file was
there before syncing. 

The sync script was also modified to make it easier to use with the release
over release syncing. This allows us to reuse it and it will also make it
easier to sync graphs manually in the case that they failed the last night. 

If the graphs are out of date, they will now display red text below the title
to indicate that it's old data (or even that it just hasn't been updated today
yet.) Before if genGraphs failed there would be no website but we need to make
it clear that the data is not from the most recent date now since the graphs
will still be there.

util/test/perf/index.html
util/test/perf/perfgraph.js 
 - add red text below title that data is out of date

util/test/genGraphs
 - creates SUCCESS file if no errors occurred 
 - puts the run date into graphdata.js so graphs know when they're out of date

util/cron/syncPerfGraphs.py
 - check for success file before syncing
 - the main change is that instead of pulling variables from envvars they are
   now arguments to the script. This way it is easy to reuse for the release
   over release graph generation and easier to call manually (or say from a
   script that will remake the graphs and attempt to sync all of them.)

util/cron/nightly
 - changes syncPerfGraphs new format of taking variables as command line args
   instead of letting it pull the same env vars
 - for the release over release now uses the syncPerfGraphs script instead of
   manually doing an rsync 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23228 3a8e244f-b0f2-452b-bcba-4c88e055c3ca