Backing out r23234
I accidentally applied this patch to the release branch rather than
trunk due to a weekend svn switch of my repository that I forgot
about.  Backing it out since it didn't make it into the module build.
That said, we should re-promote once the release is officially out 
and the branch tagged.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/release/1.9@23235 3a8e244f-b0f2-452b-bcba-4c88e055c3ca