Remove duplicate checkboxes from perf graphs
[reviewed by kyle]

We previously had a check that didn't allow duplicate graphs in the GRAPHFILE.
We changed that recently to add the 'meta' suites like performance highlights.
Because of this we had duplicate check boxes which made it annoying if you were
looking through all the graphs or searching for one in particular. 

It also meant that our graphdata.js file was 25% larger than it needed to be
because we had duplicate information (not a big deal yet but as we start adding
annotations that could become an issue.) 

This changes to generating a graph for each .graph listed to not doing it for
duplicates. What happens now is if we have seen a graph a second time, we just
append to the list of suites that it is in. Since our 'meta' suites like
performance highlights are at the top we want our graphs that are in the same
suites to be lumped together. As a result if we see a .graph again we push it to
the end of the list so it hopefully gets generated with it's original suite and
not the meta ones. 

Slight changes for compiler performance graphs:
 - I noticed the all passes, and top 10 passes didn't have different titles. 
 - For the all passes and top 10 passes suites we really want the total time
   graph to be displayed above both of the times per pass. However with this
   change it would be above the top 10, and below the all which is not the
   desired behavior. To get around this I just made a duplicate with a slightly
   different title. It's not an ideal solution, but it doesn't add that much
   clutter either. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23245 3a8e244f-b0f2-452b-bcba-4c88e055c3ca