Noinit implementation change and multiple variable declaration bug fix [reviewed by Brad]
The previous version of noinit attached a flag to variables that should not be
initialized and then upon finding this flag in normalize, created a call to
PRIM_NO_INIT.  We decided it was better style to use an established internal
symbol and reference that instead.

Additionally, multiple variable declarations were still default initializing all
but the last variable declared.  I fixed this by propagating the initialization
back to the other variables, the exact implementation of which alterred when
using the global symbol instead of the flag, so it didn't make sense to
implement one before the other.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23246 3a8e244f-b0f2-452b-bcba-4c88e055c3ca