Get chpl__readXX() calls out of generated code (a 5-minute fix (TM))
[reviewed by Lydia]

Reviewing some generated code with Lydia today, we were seeing some calls to
chpl_readXX() for basic types in the generated output, which suprised me -- I
thought they'd disappear before codegen time.  These were due to the
following overload:

        proc chpl__readXX(x) return x;

The history here is that, to get the sync/single var initialization semantics
that we want/expect, we wrap all initializing expressions when both a type
and an initializer are provided with chpl__readXX(), but then make that
function a no-op for types that are not sync/single.

Because it's a no-op, I was expecting not to see it in the generated code,
but in retrospect that was naive.  Marking the routine as 'inline' improves
things.  Where the declaration:

  var a, b, c: int;

used to result in:


        /* ChapelSyncvar.chpl:10 */
        static int64_t chpl__readXX(int64_t x) {
          return x;
        }


          call_tmp = chpl__readXX(a);

it now simply results in:

          call_tmp = a;

resulting in a minor reduction in generated code size and complexity.


While here, I also made the sync/single versions inline since they
essentially just forward to the readXX methods.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23249 3a8e244f-b0f2-452b-bcba-4c88e055c3ca