Removing error overload for {D}
[trivial, not reviewed]

For the purposes of the 1.9.0 release, we put in a warning about {D}
not meaning what it used to in case users had code that needed to be
brought up-to-date.  Now that the release branch has occurred, this
error message isn't necessary any longer.

These .goods reflect the current, not ideal, behavior.  Since
associative domains of domains are not yet implemented, we get an
error message to that effect currently for these tests.  Once domains
of domains are supported, these will need to be updated. (I considered
adding them as futures, but we already have a lot of futures in this
area, so it seemed redundant).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23250 3a8e244f-b0f2-452b-bcba-4c88e055c3ca