Added a performance test of Chapel's write() vs. C's printf() of "\n".
The test exposes Chapel's overhead on top of the actual work
of displaying the string.

To make the timing meaningful, the test performs write()/printf()
multiple times. To reduce the output volume, we use a short string.
This has the (unintentional) effect of emphasizing the overhead.

Since we expect the output to be redirected, due to its high volume,
there may be various startup effects.
To mitigate them, we rotate the variants 'tries' times,
starting with the one we expect to be the fastest.
However, the initial measurements did not support this concern.

Test variations - in C:
* Use printf() or puts().
* For C, wrap printf() in a loop in C or in Chapel.
* Pass "\n" to printf() from Chapel as a Chapel 'string' or 'c_string'.
* For the previous one, make it a global constant or a local constant.
* We could also do the above variations for puts() - skip that for now,
  to reduce the amount of info to deal with, and perhaps puts() is
  less interesting.

Test variations - in Chapel:
* since stdout is blocking, we also try a variant using a workaround of
  creating a non-blocking version of it - see stdoutNB.

With this commit's performance settings n=8M tries=5,
the test creates 320MB of output.
Luckily, start_test/sub_test pipes it into some temp space
so it should not affect the performance.
The .prediff then reduces it down to a couple of lines.

I set up verification in .perfkeys to ensure consistency of the
problem size and number of tries. It is OK to change those,
just need to take care of the historical performance data.

We should probably put together a "C vs. Chapel" suite in GRAPHFILES.
Then we can move this test into that suite.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23253 3a8e244f-b0f2-452b-bcba-4c88e055c3ca