Implement remaining assignment cases using assign-to-ref.
This patch completes the conversion of assignment to use the signature
  proc =(ref a, b) : void
.  The driving change is the removal of the catch-all assignment
  inline proc =(ref a, b) return b;
at line 258 in ChapelBase.chpl.  The remaining changes allow the implementation
to work correctly without it.  These remaining changes include (largest to
smallest):
 - Supplying a default implementation of assignment for external types (Brad's
  idea).
 - Added back more-specific versions of assignments that used to be covered by
  the catch-all.
 - Changed the semantics of record assignment
 - Updated tests and test results as necessary
 - Accommodated performance slips in some gasnet tests by adding .timeout files
  or reducing the problem size.


DETAILS
Index: test/classes/hilde/records/assignReturns-error.chpl
Index: test/classes/hilde/records/assignReturns-error.good
Added a new test to lock in the error message when the return type of an
assignop is not void.

Index: test/classes/hilde/records/assignReturns-noerror.chpl
Index: test/classes/hilde/records/assignReturns-noerror.good
Index: test/classes/hilde/records/assignReturns-noerror.future
This is a new future because even though returning a value from an assignop
should be an error, returning a bool succeeds.

Index: test/classes/bradc/records/assignRecord3.future
This test was converted into a future, because the new record assignment
semantics cause it to fail.

Index: test/classes/bradc/syncAsClass3.future
Index: test/classes/bradc/syncAsClass3.good
Index: test/classes/bradc/syncAsClass3.chpl
Added an error message future about sugaring the _syncvar and _singlevar type
names when they are printed out in error messages.

Index: test/classes/bradc/syncAsClass2.good
Removing the catch-all assignment forces assignment from a syncvar to an object
to attempt to bind to class assignment.  Since there is no coercion from
_syncvar(int(64)) to object, resolution of the assignment fails.
Previously, the catch-all assignment succeeded, and then the type mismatch was
reported in resolveMove() when resolving the wrapper PRIM_MOVE.

Index: test/types/type_variables/deitz/part7/test_tuplelike1.chpl
This test case failed due to the change in record assignment semantics.
However, providing an explicit implementation for the assignment on line 16 to succeed.
 
Index: test/types/imag/opequals2.chpl
Index: test/types/imag/opequals2.compopts
Index: test/types/imag/opequals2.good (deleted)
Index: test/types/imag/opequals2.1.good
Index: test/types/imag/opequals2.2.good
Index: test/types/imag/opequals2.3.good
Index: test/types/imag/opequals2.4.good
This test was also affected by the removal of the catch-all assignment.
Resolution of the assignment itself used to succeed, and then the error was
reported in resolveMove() when resolving the wrapper MOVE.  In that case, it was
possible to convert the fatal resolution error into a warning and have the
compiler proceed.  However, resolution of the assignment itself must succeed or
other compiler errors crop up downstream.

Taking that into consideration, the test was rewitten so that each of the errors
is evoked using a different config param instead of relying on
--ignore-errors-for-pass.

Index: test/compflags/bradc/minimalModules/minModFnRefArg.chpl
This test had to be updated because the definition for catch-all assignment is
no longer included in the minimal-module version of ChapelBase.chpl.

Index: test/expressions/hilde/assignopLHSargIsRef.chpl
Index: test/expressions/hilde/assignopLHSargIsRef.good
A new test was added to lock in the check that the LHS of an assignop has 'ref'
intent.

Index: spec/Records.tex
The spec chapter on records was updated to reflect the new record assignment
semantics.  
It used to be that assignment between arbitrary record types was valid, so long
as field-by-field assignment (by name) was valid -- the enumeration being
driven by the destination type.  In the new semantics, the source record type
must be a subtype of the destination record type.
The paragraph on assignment from a record to a variable of class type was
removed, since it is no longer pertinent (and belonged in the chapter on
classes anyway).
The section on assigning objects to records was also removed.  Since a class
type cannot be a subclass of a record type, this kind of assignment is not
supported under the new semantics.

Index: test/compflags/bradc/minimalModules/minModFnRefArg.chpl
Users of the minimal modules must now add any assignment operator that is not
internally generated by the compiler.

Index: modules-minimal/internal/ChapelBase.chpl
The catch-all assignment has been removed.  Code relying on assignments defined
in the full (as opposed to minimal) modules will have to supply those
assignments explicitly.

Index: modules/standard/List.chpl
After updating the semantics of record assignment, the special case to flag an
assignment from nil to List is no longer needed.  In general, assignment from
nil to a record type is disallowed.

Index: modules/standard/SysBasic.chpl
Index: modules/internal/ChapelTaskTable.chpl
After elimination of the catch-all for assignment, support for assignment to a
c_ptr and chpl_taskID_t had to be added explicitly.

Index: modules/internal/ChapelBase.chpl
The catch-all assignment overload was removed.
An overload was added to handle the assignment of nil to a variable of class
type.  This assignment was formerly covered by the catch-all assignment.
It was necessary to add an explicit implementation for assignment of a _ddata
class.  The question remains why such assignment was not covered by the general
overload for class assignment.

Index: compiler/AST/expr.cpp
The type of the value 'nil' is not packaged in a _ref() structure, so segments
of code that depend on this representation will mistake it for a value rather
than a reference.  In codegen, 'nil' has to be treated like literal value.  Thus
in remote puts, it has to be copied into a temporary first, and then the address
of the temporary passed as a the local buffer address in the call to
chpl_gen_comm_put().  The special casing in codegenValuePtr() takes care of
this.
To see the failure, revert the change and run
test/extern/ferguson/readme_examples.chpl with --no-local.

Index: compiler/AST/type.cpp
Index: compiler/passes/buildDefaultFunctions.cpp
Since the catch-all assignment overload was removed, the compiler must generate
default assignment implementations for external types.  This, of course,
requires that such types be properly marked with FLAG_EXTERN, including:
  dtString
  dtFile
  dtOpaque
  dtTaskID
  dtTaskList
 
Index: compiler/passes/checkResolved.cpp
Added a check that the return type of an assignment or assignop function is
void.

Index: compiler/passes/buildDefaultFunctions.cpp
Added comments and whitespace, and rearranged buildDefaultFunctions() to make
its purpose clearer.  Added the clause that calls
build_extern_assignment_function().
The RHS argument of the default record assignment function is now constrained to
be a subtype of the LHS type.  This implements the new record assignment
semantics.
Added FLAG_INLINE in the enum_comparison, enum_assignment, record_assignment,
record_cast, readThis, writeThis and extern assignment functions.

[tested on / (std)]
[tested on / (no-local)]
[reviewed by bradc]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23255 3a8e244f-b0f2-452b-bcba-4c88e055c3ca