Skip minimal modules tests for non-flat locale models
[OK'd by brad]

non-flat locale models are not supported by minimal modules since the locale
model modules are not defined. This simply adds a skipif if the local model is
not flat. 

Note that minimal modules test are currently failing for qthreads as well but
the qthreads and numa failures are unrelated (numa is failing for additional
reasons beyond the fact that it uses qthreads for the tasking layer.) 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23256 3a8e244f-b0f2-452b-bcba-4c88e055c3ca