Change log button warning behavior
[reviewed by lydia] 

Previously genGraphs looked for any values in a graphs data set that were less
than or equal to zero and if there were any it set an unloggable boolean. The
log button was then made red to warn users of potential rendering errors if
they tried to take the log of such a value. 

It turns out that the actual problem is that dygraphs screws up rendering if
the range being displayed includes zero regardless as to if there are actually
any data points there. If you do not explicitly zoom in, dygraphs auto-adjusts
the range being displayed when you take the log. So if you have the range 0-5
being displayed and the lowest point is 1, dygraphs zooms to 1 and then takes
the log. Even if you have a lowest data point of 0 and 0-5 being displayed
dygraphs will zoom the data point just above zero and will simply not render
data points at 0. However if you explicitly zoom, dygraphs uses that zoom and
does not automatically adjust so it will fail to render. So if you explicitly
zoom to 0-5 it will fail to render even if there are data points at zero. 

It was hard to try and fix this issue (you can't just automatically adjust the
zoom for a user if they have zoomed in on zero because that screws up dygraphs
trying to reset the zoom if you double click on the graph for some reason.)
Instead I just modified the existing behavior to turn the button red if they
have zoomed in on an area <= 0. 

This also adds a top margin to the buttons so that they aren't right next to
the dates.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23264 3a8e244f-b0f2-452b-bcba-4c88e055c3ca