Remove wrapper move from assignment.
Removed fix_user_assign() from normalize.cpp.  That code used to insert the
"wrapper move" that wrote the result of an assignment back to the LHS
argument.  After "ref" intents were introduced, it became possible to update
the LHS argument directly, and after all assignments were rewritten to use
"ref" intent for the LHS argument, the wrapper move became dead code.  That
means that insertion of those wrapper moves is also pointless, so at last it
can be removed as well.

A check was added in checks.cpp to ensure that no wrapper moves remain in the
tree.  This can help to isolate the cause of potential regressions associated
with this change.  If none occur (or after all have been corrected), this check
can be removed.

[tested on / (std)]
[tested on / (--verify)]
[reviewed by diten]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23266 3a8e244f-b0f2-452b-bcba-4c88e055c3ca