Hide annotations button for now
[trivial now reviewed]

Since we don't actually have any annotations yet, hiding the button as to not
confuse users. Also changed pushstate to replaceState. That way we don't add a
bunch of states to the users history but just modify the current one. (The
previous way was annoying if you use the back button.) 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23267 3a8e244f-b0f2-452b-bcba-4c88e055c3ca