Add ability to have locale model .good files
[reviewed by sung] 

We have some tests whose output is dependent on the locale model. This adds the
ability to create .good files based on the locale model. It also adds a comm
and locale model .good file option. We have some tests that have a
test.comm-none.good and a test.good file where the .good is for --no-local and
gasnet testing. Some of the numa cases are identical to the --no-local or
gasnet ones so now you can have test.comm-none.lm-flat.good. There might also
be cases where the output is different for gasnet and numa than for other
variations. The order is comm then locale model and mimics the order we use for
the regression mails. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23269 3a8e244f-b0f2-452b-bcba-4c88e055c3ca