Retired the future and moved the test where-clause-bug-1.chpl.
The future was due to a compiler crash, which is now gone.

The .good file assumed that associative domains over rectangular arrays
were allowed. We ruled that out in r20832, I believe.
So now this test correctly generates an error.

I noticed that the test added in r20832 does not test an error
that this test generates, so I moved this test to that test,
and renamed to my taste.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23270 3a8e244f-b0f2-452b-bcba-4c88e055c3ca