Added type requirements on procedures' formals in the spec.
[reviewed by Tom and discussed with Brad]

These are standard, based on what each intent does.

I did not specify them for 'const in' and 'const ref'
because they are declared "identical" to 'in' and
'ref', resp., with a difference that's unrelated
to type requirements.

Tom and I view the semantics of the 'out' intent as performing an
assignment from the formal to the actual upon the function return.
I suspect that this is not how it is currently implemented,
so I filed a future as well. From .future:

This follows up on a discussion I had with Brad and Tom, whose outcome
is documented in "the out intent" section of the spec as of this commit.

Currently the compiler implements 'out' argument passing
by reference, so the exact type match is required. However, it is
semantically possible to relax that requirement to allow
implicit conversion, or better yet, assignment from the formal
to the actual upon the function's return. Tom was excited
about this generalization, while Brad was OK with either choice.

One implementation - or performance - challenge for the more general case
seems to be a need in a temporary to store the final value of the formal
before assigning it to the actual, if their types differ.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23271 3a8e244f-b0f2-452b-bcba-4c88e055c3ca