Refine the use of FLAG_INLINE with generated functions.
Removed the "inline" flags from the
  record assignment
  union assignment
  record hash
functions, because these enumerate the record or union fields, and so are not
necessarily small.

The flag was reapplied to record assignment for an extern record, since the
body of that function is simple.  Likewise, it was reapplied to the record hash
function for a zero-length record.

[smoke-tested on hello.chpl]
[not reviewed]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23274 3a8e244f-b0f2-452b-bcba-4c88e055c3ca