Ensure that ambiguous resolution errors are tagged as ambiguous.
In printResolutionError, there are a number of special cases to interpret
resolution errors for the user.  With respect to assignment, the existing
implementation can be confusing, since failure due to multiple candidates
cannot be distinguished from failure due to the lack of any.

This patch splits the former implementation into two:
  printResolutionErrorAmbiguous
  printResolutionErrorUnresolved
.  As it turns out, there was not a lot of overlap between them.  Some
factoring of common code could be performed at finer granularity; that is left
as a future task.

When the call is unresolved, the observed behavior will be the same as the
current implementation.  When the call is ambiguous, however, the new
implementation ensures that it is diagnosed as ambiguous.

Three test cases were contributed by Vass.

[tested on / (std)]
[reviewed by vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23277 3a8e244f-b0f2-452b-bcba-4c88e055c3ca