Fix numa dataParIgnoreRunning output regression
[reviewed by sung]

test/studies/hpcc/common/bradc/unit/multipleDomains2.chpl prints out the
dataParIgnoreRunning status. It is true for most configurations but false for
numa. All this patch does is change the comm-none.good to
comm-none.lm-flat.good and creates a new lm-numa.good file that is identical to
the regular .good file with the execption of the two dataParIgnoreRunning
prints. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23279 3a8e244f-b0f2-452b-bcba-4c88e055c3ca