Updates to tests that check the behavior of chpldev_refToString for numa
[reviewed by david] 

There were two tests that were checking the behavior of the chpldev_refToString
method. This method prints out something like 0xNNNNNNNN for the narrow case and
0:0xNNNNNNNN for the wide case. 

trivial/bradc/ref2string explicitly threw --local and only had .good files for
the narrow case. However for numa, even when local, there can be wide strings so
there needs to be a .good for numa. Instead of just adding a numa .good file I
removed the .compopts that threw --local and added .goods to handle all wide
and non wide cases. 

For trivial/diten/classRef2string-local that problem was the same as for the
bradc/ref2String test case. However there was also a
trivial/diten/classRef2string-no-local test as well. I combined these tests
into one, removed the .compopts that threw --local and --no-local and added the
.good files to handle all wide and non wide cases. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23280 3a8e244f-b0f2-452b-bcba-4c88e055c3ca