Fix test ressults in classes/hilde/records due to r23266 (remove wrapper moves).
assignReturns-noerror now passes, so I removed the .bad and .future files and
renamed the .chpl and .good files using the base name assignReturns-error2.

Added a SKIPIF to skip these tests because they fail when --verify is tossed.
Returning a value from an assignment causes its return type to be non-void.
Then insertReturnTemps() inserts a temporary to capture this.  That construct
has the same form as a wrapper move, so gets flagged by the
checkNoWrapperMoves() check (and that changes the error message).

[reviewed by diten]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23281 3a8e244f-b0f2-452b-bcba-4c88e055c3ca