Change .good files for "on statement in iterator" tests to fix numa regressions
[reviewed by sung] 

Updates the .good files for the following tests: 
  test/functions/iterators/bradc/onInIterator2.chpl
  test/users/shetag/iteratorWithOn.chpl 

On statements inside an iterator are illegal for non local cases. These two
tests ensure that if there is an on inside an iterator an error message occurs.
Currently the error message does not occur for local compilation since on
statements are dropped in parsing. For numa, these tests should also produce
the error, which they do, but the wrong .good file was being used. This just
changes the .comm-none.good files to .comm-none.lm-flat.good so that for numa
the normal .good files with the expected error in them are used instead of
.comm-none.good. 

I also deleted the onInIterator2.skipif because it is not needed. There was
already a --no-local.good so there is no reason to skip it. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23285 3a8e244f-b0f2-452b-bcba-4c88e055c3ca